+++ CHANGELOG.md
@@ -23,6 +23,8 @@
  * feat(cli:sdk): Added aliases to sdk commands (i, ls, rm).
  * feat(sdk): Added `find` endpoint to SDK service to get info about an installed Titanium SDK.
  * feat(sdk): Added progress bars during SDK installation.
+ * feat(module): Added `/module/check-downloads` endpoint.
+ * feat(module): Added automatic checking of new Titanium module downloads.
  * feat: Support for Titanium-specific telemetry.
  * refactor: Updated to latest cli-kit with support for the new client/server architecture.
  * refactor: Updated `config` command actions to be subcommands with improved help output.